Refinements
We now have a better buffering strategy in place for the media controller. If the track fits in the buffer, we just buffer it completely asap (loop controlled by the buffer-update event).

If it doesn't fit then we incrementally update the buffer, with the loop controlled by the timeupdate-event. By default we buffer 1 min of data ahead, and remove trailing data more than a 1 min behind the current playtime.