Switch to stream reader instead of byte range-requests
Massive changes, too many to list. Mostly the change is about swiching the StreamController to using the ReadableStream API instead of relying on repeated byte-range requests.

The rest of the code is changing as well, getting cleaned up and whipped into shape so we can actually use this for a real project going forward.