Bits and bobs (byte range requests)
Switched to using byte range requests for streaming instead of StreamReader. Turns out the stream reader keeps downloading in the background which is not what we want.

CSS has finally been put in its own file.

Re-reworked the selection of tracks from the library so that playlist.add() or remove() is only called once whether you select everything across the library, artist or album. I never liked the idea of hitting remove() for each album which would re-allocate the list-array each time.

Added font-awesome to the project.