Playlist select/delete fix
Fixed selecting and deleting tracks from the playlist. It wasn't updating the selected tracks.

This also made me change the behaviour of _updateAllSelectionCounts() so it now wipes the selection counts first, otherwise it may end up doubling the count.

Put it some code to catch and properly info-log when the end of the playlist is reached.

Renamed the MEDIA_CONTROLLER_TRACK_ROTATED-event to a more verbose log message.